style modify

+++ src/backend/cpu_backend_pack-inl.h
@@ -12,23 +12,23 @@ void Backend<CPUTensor, DType>::Unpack2DParallelFunc(
   DType* unpack) {
   // (input_channel * filter_height * filter_width) *
   // (output_width * output_height)
-  int unpack_index;
-  int channel_offset;
+  size_t unpack_index;
+  size_t channel_offset;
   size_t filter_channel_offset = filter_height * filter_width;
   size_t output_channel_offset = output_height * output_width;
   size_t input_channel_offset = input_height * input_width;
   #pragma omp parallel for collapse(2) private(channel_offset, unpack_index)
-  for (int channel_index = 0; channel_index < channel; ++channel_index) {
-    for (int filter_height_index = 0; filter_height_index < filter_height;
+  for (size_t channel_index = 0; channel_index < channel; ++channel_index) {
+    for (size_t filter_height_index = 0; filter_height_index < filter_height;
         ++filter_height_index) {
       channel_offset = channel_index * input_channel_offset;
       unpack_index = (channel_index * filter_channel_offset +
         filter_height_index * filter_width) * output_channel_offset;
       const DType* input_slice = input + channel_offset;
-      for (int filter_width_index = 0; filter_width_index < filter_width;
+      for (size_t filter_width_index = 0; filter_width_index < filter_width;
           ++filter_width_index) {
-        int filter_height_offset = -padding_height + filter_height_index;
-        for (int output_height_index = 0; output_height_index < output_height;
+        size_t filter_height_offset = -padding_height + filter_height_index;
+        for (size_t output_height_index = 0; output_height_index < output_height;
             ++output_height_index) {
           if (filter_height_offset < 0 || filter_height_offset >= input_height) {
             for (int output_width_index = 0; output_width_index < output_width;
@@ -36,8 +36,8 @@ void Backend<CPUTensor, DType>::Unpack2DParallelFunc(
               unpack[unpack_index++] = 0;
             }
           } else {
-            int filter_width_offset = -padding_width + filter_width_index;
-            for (int output_width_index = 0; output_width_index < output_width;
+            size_t filter_width_offset = -padding_width + filter_width_index;
+            for (size_t output_width_index = 0; output_width_index < output_width;
                 ++output_width_index) {
               if (filter_width_offset < 0 || filter_width_offset >= input_width) {
                 unpack[unpack_index++] = 0;
@@ -73,23 +73,23 @@ void Backend<CPUTensor, DType>::Pack2DParallelFunc(
   size_t output_channel_offset = output_height * output_width;
   size_t input_channel_offset = input_height * input_width;
   #pragma omp parallel for collapse(2) private(channel_offset, pack_index)
-  for (int channel_index = 0; channel_index < channel; ++channel_index) {
-    for (int filter_height_index = 0; filter_height_index < filter_height;
+  for (size_t channel_index = 0; channel_index < channel; ++channel_index) {
+    for (size_t filter_height_index = 0; filter_height_index < filter_height;
         ++filter_height_index) {
       channel_offset = channel_index * input_channel_offset;
       pack_index = (channel_index * filter_channel_offset +
         filter_height_index * filter_width) * output_channel_offset;
       DType* input_slice = input + channel_offset;
-      for (int filter_width_index = 0; filter_width_index < filter_width;
+      for (size_t filter_width_index = 0; filter_width_index < filter_width;
           ++filter_width_index) {
-        int filter_height_offset = -padding_height + filter_height_index;
-        for (int output_height_index = 0; output_height_index < output_height;
+        size_t filter_height_offset = -padding_height + filter_height_index;
+        for (size_t output_height_index = 0; output_height_index < output_height;
             ++output_height_index) {
           if (filter_height_offset < 0 || filter_height_offset >= input_height) {
             pack_index += output_width;
           } else {
-            int filter_width_offset = -padding_width + filter_width_index;
-            for (int output_width_index = 0; output_width_index < output_width;
+            size_t filter_width_offset = -padding_width + filter_width_index;
+            for (size_t output_width_index = 0; output_width_index < output_width;
                 ++output_width_index) {
               if (filter_width_offset >= 0 && filter_width_offset < input_width) {
                 #pragma omp atomic
@@ -118,27 +118,27 @@ void Backend<CPUTensor, DType>::Unpack2DFunc(
   DType* unpack) {
   // (input_channel * filter_height * filter_width) *
   // (output_width * output_height)
-  int unpack_index = 0;
-  int channel_offset;
+  size_t unpack_index = 0;
+  size_t channel_offset;
   size_t input_channel_offset = input_height * input_width;
-  for (int channel_index = 0; channel_index < channel; ++channel_index) {
+  for (size_t channel_index = 0; channel_index < channel; ++channel_index) {
     channel_offset = channel_index * input_channel_offset;
     const DType* input_slice = input + channel_offset;
-    for (int filter_height_index = 0; filter_height_index < filter_height;
+    for (size_t filter_height_index = 0; filter_height_index < filter_height;
         ++filter_height_index) {
-      for (int filter_width_index = 0; filter_width_index < filter_width;
+      for (size_t filter_width_index = 0; filter_width_index < filter_width;
           ++filter_width_index) {
-        int filter_height_offset = -padding_height + filter_height_index;
-        for (int output_height_index = 0; output_height_index < output_height;
+        size_t filter_height_offset = -padding_height + filter_height_index;
+        for (size_t output_height_index = 0; output_height_index < output_height;
             ++output_height_index) {
           if (filter_height_offset < 0 || filter_height_offset >= input_height) {
-            for (int output_width_index = 0; output_width_index < output_width;
+            for (size_t output_width_index = 0; output_width_index < output_width;
                 ++output_width_index) {
               unpack[unpack_index++] = 0;
             }
           } else {
-            int filter_width_offset = -padding_width + filter_width_index;
-            for (int output_width_index = 0; output_width_index < output_width;
+            size_t filter_width_offset = -padding_width + filter_width_index;
+            for (size_t output_width_index = 0; output_width_index < output_width;
                 ++output_width_index) {
               if (filter_width_offset < 0 || filter_width_offset >= input_width) {
                 unpack[unpack_index++] = 0;
@@ -165,24 +165,24 @@ void Backend<CPUTensor, DType>::Pack2DFunc(
   size_t padding_height, size_t padding_width,
   size_t stride_height, size_t stride_width,
   DType* input) {
-  int pack_index = 0;
-  int channel_offset;
+  size_t pack_index = 0;
+  size_t channel_offset;
   size_t input_channel_offset = input_height * input_width;
-  for (int channel_index = 0; channel_index < channel; ++channel_index) {
+  for (size_t channel_index = 0; channel_index < channel; ++channel_index) {
     channel_offset = channel_index * input_channel_offset;
     DType* input_slice = input + channel_offset;
-    for (int filter_height_index = 0; filter_height_index < filter_height;
+    for (size_t filter_height_index = 0; filter_height_index < filter_height;
         ++filter_height_index) {
-      for (int filter_width_index = 0; filter_width_index < filter_width;
+      for (size_t filter_width_index = 0; filter_width_index < filter_width;
           ++filter_width_index) {
         int filter_height_offset = -padding_height + filter_height_index;
-        for (int output_height_index = 0; output_height_index < output_height;
+        for (size_t output_height_index = 0; output_height_index < output_height;
             ++output_height_index) {
           if (filter_height_offset < 0 || filter_height_offset >= input_height) {
             pack_index += output_width;
           } else {
-            int filter_width_offset = -padding_width + filter_width_index;
-            for (int output_width_index = 0; output_width_index < output_width;
+            size_t filter_width_offset = -padding_width + filter_width_index;
+            for (size_t output_width_index = 0; output_width_index < output_width;
                 ++output_width_index) {
               if (filter_width_offset >= 0 && filter_width_offset < input_width) {
                 input_slice[filter_height_offset * input_width +
