+++ src/backend/cpu_backend_pool-inl.h
@@ -10,15 +10,15 @@ void Backend<CPUTensor, DType>::MaxPooling2DForwardFunc(
   // shape decode
   // input
   const Shape& input_shape = input->shape();
-  int batch_size = input_shape[0];
-  int input_channel = input_shape[1];
-  int input_height = input_shape[2];
-  int input_width = input_shape[3];
+  size_t batch_size = input_shape[0];
+  size_t input_channel = input_shape[1];
+  size_t input_height = input_shape[2];
+  size_t input_width = input_shape[3];
   // output
   const Shape& output_shape = output->shape();
-  int output_channel = output_shape[1];
-  int output_height = output_shape[2];
-  int output_width = output_shape[3];
+  size_t output_channel = output_shape[1];
+  size_t output_height = output_shape[2];
+  size_t output_width = output_shape[3];
 
   CHECK_EQ(input_channel, output_channel);
 
@@ -34,10 +34,10 @@ void Backend<CPUTensor, DType>::MaxPooling2DForwardFunc(
   // no padding
   #pragma omp parallel for private(batch_input_offset, batch_output_offset, \
     channel_input_offset, channel_output_offset)
-  for (int batch_index = 0; batch_index < batch_size; ++batch_index) {
+  for (size_t batch_index = 0; batch_index < batch_size; ++batch_index) {
     batch_input_offset = batch_index * input_single_offset;
     batch_output_offset = batch_index * output_single_offset;
-    for (int channel_index = 0; channel_index < input_channel; ++channel_index) {
+    for (size_t channel_index = 0; channel_index < input_channel; ++channel_index) {
       channel_input_offset = channel_index * input_channel_offset;
       channel_output_offset = channel_index * output_channel_offset;
       const DType* input_slice = input->Slice(batch_input_offset +
@@ -46,9 +46,9 @@ void Backend<CPUTensor, DType>::MaxPooling2DForwardFunc(
         channel_output_offset);
       size_t* max_index_slice = max_index->Slice(batch_output_offset +
         channel_output_offset);
-      for (int output_height_index = 0; output_height_index < output_height;
+      for (size_t output_height_index = 0; output_height_index < output_height;
         ++output_height_index) {
-        for (int output_width_index = 0; output_width_index < output_width;
+        for (size_t output_width_index = 0; output_width_index < output_width;
           ++output_width_index) {
           size_t height_start = output_height_index * stride_height;
           size_t width_start = output_width_index * stride_width;
@@ -56,7 +56,7 @@ void Backend<CPUTensor, DType>::MaxPooling2DForwardFunc(
           size_t width_end = width_start + filter_width;
           size_t pool_index = output_height_index * output_width +
             output_width_index;
-          int max_index_tmp = height_start * input_width + width_start;
+          size_t max_index_tmp = height_start * input_width + width_start;
           for (int h = height_start; h < height_end; ++h) {
             for (int w = width_start; w < width_end; ++w) {
               size_t index = h * input_width + w;
@@ -82,14 +82,14 @@ void Backend<CPUTensor, DType>::MaxPooling2DBackwardFunc(
   // shape decode
   // input
   const Shape& input_shape = input->shape();
-  int batch_size = input_shape[0];
-  int channel = input_shape[1];
-  int input_height = input_shape[2];
-  int input_width = input_shape[3];
+  size_t batch_size = input_shape[0];
+  size_t channel = input_shape[1];
+  size_t input_height = input_shape[2];
+  size_t input_width = input_shape[3];
   // output
   const Shape& output_shape = output->shape();
-  int output_height = output_shape[2];
-  int output_width = output_shape[3];
+  size_t output_height = output_shape[2];
+  size_t output_width = output_shape[3];
 
   size_t input_single_offset = channel * input_height * input_width;
   size_t input_channel_offset = input_height * input_width;
@@ -104,10 +104,10 @@ void Backend<CPUTensor, DType>::MaxPooling2DBackwardFunc(
   // no padding
   #pragma omp parallel for private(batch_input_offset, batch_output_offset, \
     channel_input_offset, channel_output_offset)
-  for (int batch_index = 0; batch_index < batch_size; ++batch_index) {
+  for (size_t batch_index = 0; batch_index < batch_size; ++batch_index) {
     batch_input_offset = batch_index * input_single_offset;
     batch_output_offset = batch_index * output_single_offset;
-    for (int channel_index = 0; channel_index < channel; ++channel_index) {
+    for (size_t channel_index = 0; channel_index < channel; ++channel_index) {
       channel_input_offset = channel_index * input_channel_offset;
       channel_output_offset = channel_index * output_channel_offset;
       DType* input_slice = input->Slice(batch_input_offset +
@@ -116,9 +116,9 @@ void Backend<CPUTensor, DType>::MaxPooling2DBackwardFunc(
         channel_output_offset);
       const size_t* max_index_slice = max_index->Slice(batch_output_offset +
         channel_output_offset);
-      for (int output_height_index = 0; output_height_index < output_height;
+      for (size_t output_height_index = 0; output_height_index < output_height;
         ++output_height_index) {
-        for (int output_width_index = 0; output_width_index < output_width;
+        for (size_t output_width_index = 0; output_width_index < output_width;
           ++output_width_index) {
           size_t index = output_height_index * output_width +
             output_width_index;
